Allow local copy of mod_cfml jar

+++ scripts/300-tomcat.sh
@@ -24,16 +24,22 @@ cp /etc/tomcat7/catalina.properties backup/etc/tomcat7/catalina.properties-orig-
 #copy our catalina properties
 cp etc/tomcat7/catalina.properties /etc/tomcat7/
 
+cp /etc/default/tomcat7 backup/etc/default/tomcat7
+
 echo "Installing Railo Java Agent into Tomcat"
 echo $'\nJAVA_OPTS="${JAVA_OPTS} -javaagent:/opt/railo/current/railo-inst.jar"' >> /etc/default/tomcat7
 
 echo "Installing mod_cfml Valve for Automatic Virtual Host Configuration"
-curl -o /opt/railo/current/mod_cfml-valve_v1.0.14.jar http://download.modcfml.org/downloader.cfm/id/24/file/mod_cfml-valve_v1.0.14.jar
-
+if [ -f lib/mod_cfml-valve_v1.0.14.jar ]; then
+  cp lib/mod_cfml-valve_v1.0.14.jar /opt/railo/current/
+else
+  curl -o /opt/railo/current/mod_cfml-valve_v1.0.14.jar http://download.modcfml.org/downloader.cfm/id/24/file/mod_cfml-valve_v1.0.14.jar
+fi
+   
 echo "Setting Permissions on Railo Folders"
 chown -R tomcat7:tomcat7 /opt/railo
 chmod -R 750 /opt/railo
 
 echo "Setting JVM Max Heap Size to " $JVM_MAX_HEAP_SIZE
-cp /etc/default/tomcat7 backup/etc/default/tomcat7
+
 sed -e "s/-Xmx128m/-Xmx$JVM_MAX_HEAP_SIZE/g" /etc/default/tomcat7 >> /etc/default/tomcat7
